requirements.txt
torch>=2.4.0
transformers>=4.46.0
datasets>=3.0.0
accelerate>=1.0.0
peft>=0.13.0
trl>=0.12.0
bitsandbytes>=0.44.0
flash-attn>=2.6.0
sentencepiece>=0.2.0
protobuf>=4.25.0
scipy>=1.13.0
pyyaml>=6.0
wandb>=0.18.0
huggingface_hub>=0.26.0
safetensors>=0.4.0

# Data collection & scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Synthetic data generation (choose one or both)
openai>=1.12.0
anthropic>=0.25.0
